add climate_functions.py; move related functions to it; fix extract_climate_data by adding _ to hist and future
