+++ climate_functions.py
@@ -1,0 +1,114 @@
+"""
+Functions for climat data extracting, processing, analysis.
+climate patterns, historical and future climate projections
+"""
+import streamlit as st
+import xarray as xr
+import numpy as np
+import pandas as pd
+
+#from climate_functions import (
+#    load_data,
+#    extract_climate_data
+#)
+
+@st.cache_data
+def load_data(data_path):
+    """
+    load climate model data 
+
+    Args:
+        data_path (str): path to the data with model runs (historical and projection)
+
+    Returns:
+        xarray.core.dataset.Dataset, xarray.core.dataset.Dataset : data from historical (hindacst) runs and climate projection
+    """    
+    hist = xr.open_mfdataset(f"{data_path}/AWI_CM_mm_historical*.nc", compat="override")
+    future = xr.open_mfdataset(f"{data_path}/AWI_CM_mm_ssp585*.nc", compat="override")
+    return hist, future
+
+def convert_to_mm_per_month(monthly_precip_kg_m2_s1):
+    days_in_months = np.array([31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31])
+    return monthly_precip_kg_m2_s1 * 60 * 60 * 24 * days_in_months
+
+
+@st.cache_data
+def extract_climate_data(lat, lon, _hist, _future):
+    """
+    Extracts climate data for a given latitude and longitude from historical and future datasets.
+
+    Args:
+    - lat (float): Latitude of the location to extract data for.
+    - lon (float): Longitude of the location to extract data for.
+    - hist (xarray.Dataset): Historical climate dataset.
+    - future (xarray.Dataset): Future climate dataset.
+
+    Returns:
+    - df (pandas.DataFrame): DataFrame containing present day and future temperature, precipitation, and wind speed data for each month of the year.
+    - data_dict (dict): Dictionary containing string representations of the extracted climate data.
+    """
+    hist_temp = _hist.sel(lat=lat, lon=lon, method="nearest")["tas"].values - 273.15
+    hist_temp_str = np.array2string(hist_temp.ravel(), precision=3, max_line_width=100)[
+        1:-1
+    ]
+
+    hist_pr = _hist.sel(lat=lat, lon=lon, method="nearest")["pr"].values
+    hist_pr = convert_to_mm_per_month(hist_pr)
+
+    hist_pr_str = np.array2string(hist_pr.ravel(), precision=3, max_line_width=100)[
+        1:-1
+    ]
+
+    hist_uas = _hist.sel(lat=lat, lon=lon, method="nearest")["uas"].values
+    hist_uas_str = np.array2string(hist_uas.ravel(), precision=3, max_line_width=100)[
+        1:-1
+    ]
+
+    hist_vas = _hist.sel(lat=lat, lon=lon, method="nearest")["vas"].values
+    hist_vas_str = np.array2string(hist_vas.ravel(), precision=3, max_line_width=100)[
+        1:-1
+    ]
+
+    future_temp = _future.sel(lat=lat, lon=lon, method="nearest")["tas"].values - 273.15
+    future_temp_str = np.array2string(
+        future_temp.ravel(), precision=3, max_line_width=100
+    )[1:-1]
+
+    future_pr = _future.sel(lat=lat, lon=lon, method="nearest")["pr"].values
+    future_pr = convert_to_mm_per_month(future_pr)
+    future_pr_str = np.array2string(future_pr.ravel(), precision=3, max_line_width=100)[
+        1:-1
+    ]
+
+    future_uas = _future.sel(lat=lat, lon=lon, method="nearest")["uas"].values
+    future_uas_str = np.array2string(
+        future_uas.ravel(), precision=3, max_line_width=100
+    )[1:-1]
+
+    future_vas = _future.sel(lat=lat, lon=lon, method="nearest")["vas"].values
+    future_vas_str = np.array2string(
+        future_vas.ravel(), precision=3, max_line_width=100
+    )[1:-1]
+    df = pd.DataFrame(
+        {
+            "Present Day Temperature": hist_temp[0, 0, :],
+            "Future Temperature": future_temp[0, 0, :],
+            "Present Day Precipitation": hist_pr[0, 0, :],
+            "Future Precipitation": future_pr[0, 0, :],
+            "Present Day Wind Speed": np.hypot(hist_uas[0, 0, :], hist_vas[0, 0, :]),
+            "Future Wind Speed": np.hypot(future_uas[0, 0, :], future_vas[0, 0, :]),
+            "Month": range(1, 13),
+        }
+    )
+    data_dict = {
+        "hist_temp": hist_temp_str,
+        "hist_pr": hist_pr_str,
+        "hist_uas": hist_uas_str,
+        "hist_vas": hist_vas_str,
+        "future_temp": future_temp_str,
+        "future_pr": future_pr_str,
+        "future_uas": future_uas_str,
+        "future_vas": future_vas_str,
+    }
+    return df, data_dict
+
